+++ sphinxcontrib/needs/needs.py
@@ -517,19 +517,17 @@ def check_configuration(_app: Sphinx, config: Config):
     link_types = [x["option"] for x in config["needs_extra_links"]]
 
     log = get_logger(__name__)
-    external_filter = config["needs_filter_data"]
-    # Check if external filter values is really a string
+    external_filter = getattr(config, "needs_filter_data", {})
     for extern_filter, value in external_filter.items():
+        # Check if external filter values is really a string
         if not isinstance(value, str):
             log.warning(
                 "External filter value: {0} from needs_filter_data {1} is not a string.".format(value, external_filter)
             )
-
-    # Check if needs external filter and extra option are using the same name
-    for ext_filter in external_filter.keys():
-        if ext_filter in extra_options:
+        # Check if needs external filter and extra option are using the same name
+        if extern_filter in extra_options:
             raise NeedsConfigException(
-                "Same name for external filter and extra option: {}." " This is not allowed.".format(ext_filter)
+                "Same name for external filter and extra option: {}." " This is not allowed.".format(extern_filter)
             )
 
     # Check for usage of internal names